Update Pattern3.pde
I worked on pattern3 at home. Made a cos wave because I thought it would be cool. It was a bit challenging to get the offset to work so the lines did not all just move in unison (which would defeat the purpose of a wave) but I got it eventually.