Update sketch_2PG_Project.pde
Edited it to fix a card issue (only 31 cards were showing in the game and then I released that it was an issue with the cardX coordinates rather than the code itself - it had to be less than 600 not 700 or else it would go off the screen). I also tried adding animations (still working on it).