Update Game.pde
Finished up most of the fine tuning - added the player scoring and player turn change. There is some weird lag that goes on near the end of the game when most of the cards have already been guessed correctly. Will try to see if its a problem in my code.